requirements.txt
asgiref==3.7.2
attrs==23.1.0
certifi==2023.7.22
click==8.1.7
colorama==0.4.6
Django==4.2.2
django-debug-toolbar==4.2.0
django-filter==23.3
djangorestframework==3.14.0
dnspython==2.4.2
docutils==0.20.1
drf-spectacular==0.26.5
email-validator==2.0.0.post2
flake8==6.1.0
flake8-docstrings==1.7.0
Flask==2.2.3
Flask-WTF==1.1.1
freezegun==1.2.2
idna==3.4
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
logging-tree==1.9
MarkupSafe==2.1.3
mccabe==0.7.0
Pillow==10.0.1
product==0.1.1.1
pycodestyle==2.11.1
pydocstyle==6.3.0
pyflakes==3.1.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
rpds-py==0.10.6
sentry-sdk==1.32.0
six==1.16.0
snowballstemmer==2.2.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.7
Werkzeug==2.3.7
WTForms==3.0.1
